2017年2月28号，完成青海无参考出库，转储退库，验收结果入库-其他，采购退货，以及委外入库几个功能模块。还需要完善的是： 1.委外入库的服务端，2. 105非必检入库的测试，以及105必检入库整个模块，3. 无参考退库，盘点以及检验批验收等模块需要尽快的做完，测试。

+++ app/src/main/java/com/richfit/barcodesystemproduct/module/setting/ISettingView.java
@@ -26,7 +26,7 @@ public interface ISettingView extends BaseView {
 
     //下载基础数据
     void onStartLoadBasicData(int maxProgress);
-    void loadBasicDataProgress(int progress);
+    void loadBasicDataProgress(float progress);
     void loadBasicDataFail(String message);
     void loadBasicDataSuccess();
